requirements.txt
torch==1.11.0
Pillow==9.1.1
matplotlib==3.5.2
pytorch-lightning==1.6.4
torchvision==0.12.0
gdown==4.4.0
